requirements.txt
blue==0.9.0
isort==5.10.1
mypy==0.960
pikepdf==5.1.3
pip-chill==1.0.1
pypdf2==1.28.4
pytest-cov==3.0.0
requests==2.27.1
types-requests==2.27.29